third_party/blink/web_tests/external/wpt/css/css-transforms/parsing/transform-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['transform'] = "scaleX(7)" should set the property value
  assert_equals: serialization should be canonical expected "scalex(7)" but got "scaleX(7)"
[FAIL] e.style['transform'] = "scaleX(720%)" should set the property value
  assert_equals: serialization should be canonical expected "scalex(7.2)" but got "scaleX(7.2)"
[FAIL] e.style['transform'] = "scaleY(-8)" should set the property value
  assert_equals: serialization should be canonical expected "scaley(-8)" but got "scaleY(-8)"
[FAIL] e.style['transform'] = "scaleY(-85%)" should set the property value
  assert_equals: serialization should be canonical expected "scaley(-0.85)" but got "scaleY(-0.85)"
[FAIL] e.style['transform'] = "scaleZ(4)" should set the property value
  assert_equals: serialization should be canonical expected "scalez(4)" but got "scaleZ(4)"
[FAIL] e.style['transform'] = "scaleZ(25%)" should set the property value
  assert_equals: serialization should be canonical expected "scalez(0.25)" but got "scaleZ(0.25)"
[FAIL] e.style['transform'] = "skewX(0)" should set the property value
  assert_equals: serialization should be canonical expected "skewx(0deg)" but got "skewX(0deg)"
[FAIL] e.style['transform'] = "skewX(90deg)" should set the property value
  assert_equals: serialization should be canonical expected "skewx(90deg)" but got "skewX(90deg)"
[FAIL] e.style['transform'] = "skewY(0)" should set the property value
  assert_equals: serialization should be canonical expected "skewy(0deg)" but got "skewY(0deg)"
[FAIL] e.style['transform'] = "skewY(-90deg)" should set the property value
  assert_equals: serialization should be canonical expected "skewy(-90deg)" but got "skewY(-90deg)"
Harness: the test ran to completion.
